test: split up test suite
#163
